requirements.txt
Flask-PyMongo==2.1.0
Flask-RESTful==0.3.6
Flask-User==1.0.1.5
PyMongo==3.7.1